requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure.identity
azure.functions
microsoft-bing-newssearch
azure.core
azure.storage.blob
azure.storage.file.datalake
azure.keyvault.secrets
beautifulsoup4